Fix building warnings

+++ src/xrGame/ai/monsters/control_manager_custom.cpp
@@ -540,7 +540,7 @@ void CControlManagerCustom::check_rotation_jump()
 	SControlRotationJumpData	*ctrl_data = (SControlRotationJumpData *) m_man->data(this, ControlCom::eControlRotationJump);
 	VERIFY						(ctrl_data);
 
-	(*ctrl_data)				= m_rot_jump_data[Random.randI(m_rot_jump_data.size())];
+	(*ctrl_data)				= m_rot_jump_data[Random.randI((u32)m_rot_jump_data.size())];
 
 	m_man->activate				(ControlCom::eControlRotationJump);
 }
